requirements.txt
pyleniumio==1.20.0
pytest-rerunfailures==12.0
blinker==1.7.0
setuptools>=60.0.0
webdriver_manager

